fix(export): catch backtracking edges + split rooms module
- is_simple_polygon: also reject adjacent backtracking spikes (an edge
  that doubles back over its neighbour). The non-adjacent crossing/overlap
  loops skipped the shared-endpoint pair, so a >tol overlap along one edge
  could still pass and export a self-overlapping room. Reuse the collinear-
  overlap test on the (prev→cur, cur→next) pair (a touch point alone isn't
  an overlap, only a genuine doubling-back).
- split the geometry predicates out of rooms.rs (424 lines, over the
  AGENTS.md ~400 limit) into a new `geom` module; rooms.rs is now 210.

(Codex.) cargo test green; duplex check_all PASS.
